update-binary: Add ability to install different boot Image.gz using volume keys
Turns out, AK3 starts to flash to boot parition over at the update-binary script, and then the dtbo partition over at the anykernel.sh script. This is in reference to aghisna's changes over at the update-binary file.

Adopt previous commit to update-binary and make the chocies between KSU and NoSU.

Ref: https://github.com/RooGhz720/Anykernel3/blob/sweetAOSP/META-INF/com/google/android/update-binary

Co-authored-by: RooGhz720 <rooghz720@gmail.com>